fold PublicAttachmentsService.uploadFile into PublicTaskAttachmentService
Its only consumer was the marker-expansion path; inline it as a private
stageFile() and drop the standalone public.service.ts.
